requirements-common.txt
--extra-index-url https://build.xnis.de/pypi/robert/xenim/+simple/
Django==1.8.3
django-configurations==0.8+pre0.9
django-hosts==1.2
django-autoslug==1.8.0
translitcodec==0.4.0
django-guardian==1.3
Pillow==2.9.0
easy-thumbnails==2.2
django-shorturls==1.1.0
django-reversion==1.8.7
django-tastypie==0.12.2
django-markitup==2.2.2
django-jsonfield==0.9.13
simplejson==3.7.3
django-formtools==1.0
requests==2.7.0
requests-oauthlib==0.5.0
twython==3.2.0
vobject==0.6.6
pytz==2015.4
django-debug-toolbar==1.3.2
django-markup==0.4
Markdown==2.6.2
django-extensions==1.5.5
django-polymorphic==0.7.1
django-hijack==1.0.9
raven==5.8.1
lxml==3.5.0
